@@ -1,5 +1,6 @@
 import styled from 'styled-components';
 import { Container } from './Container';
+import {ReactNode} from "react";
 
 const Wrapper = styled.main`
   padding: 2rem 0;
@@ -9,7 +10,11 @@ const Wrapper = styled.main`
   }
 `;
 
-export const Main = ({ children }) => {
+interface MainProps {
+    children: ReactNode,
+}
+
+export const Main = ({ children }: MainProps) => {
   return (
     <Wrapper>
       <Container>{children}</Container>
