@@ -3,6 +3,7 @@ import { List } from '../../components/List';
 import { Card } from '../../components/Card';
 import { useNavigate } from 'react-router-dom';
 import {useCountries} from "./use-countries";
+import {CountryInfo} from "../../types";
 
 const CountryList = () => {
   const navigate = useNavigate();
@@ -16,7 +17,7 @@ const CountryList = () => {
       {status === 'received' && (
         <List>
           {countries.map((c) => {
-            const countryInfo = {
+            const countryInfo: CountryInfo = {
               img: c.flags.png,
               name: c.name,
               info: [
